% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.data}
\alias{as_json.data}
\title{Export a Datalist object to JSON}
\usage{
as_json.data(data, pretty = TRUE)
}
\arguments{
\item{data}{Datalist object}

\item{pretty}{Logical}
}
\value{
JSON string
}
\description{
Export a Datalist object to JSON
}
