% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.kernels}
\alias{as_json.kernels}
\title{Export kernel specifications to JSON}
\usage{
as_json.kernels(vcs, pretty = TRUE)
}
\arguments{
\item{vcs}{Named list of vc objects}

\item{pretty}{Logical}
}
\value{
JSON string
}
\description{
Export kernel specifications to JSON
}
