% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.model}
\alias{as_json.model}
\title{Export model specification to JSON}
\usage{
as_json.model(formulas, task, options = list(), pretty = TRUE)
}
\arguments{
\item{formulas}{Named list of formulas}

\item{task}{Character}

\item{options}{Optional list}

\item{pretty}{Logical}
}
\value{
JSON string
}
\description{
Export model specification to JSON
}
