% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.priors}
\alias{as_json.priors}
\title{Export prior specifications to JSON}
\usage{
as_json.priors(priors, pretty = TRUE)
}
\arguments{
\item{priors}{Named list of prior objects}

\item{pretty}{Logical}
}
\value{
JSON string
}
\description{
Export prior specifications to JSON
}
