% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.vc}
\alias{as_json.vc}
\title{Export a vc object to JSON}
\usage{
as_json.vc(x, pretty = TRUE, auto_unbox = TRUE)
}
\arguments{
\item{x}{An object of class "vc"}

\item{pretty}{Logical. Pretty-print JSON.}

\item{auto_unbox}{Logical. Unbox scalars.}
}
\value{
A JSON string
}
\description{
Export a vc object to JSON
}
