% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_json.vcs}
\alias{as_json.vcs}
\title{Export a list of variance components to JSON}
\usage{
as_json.vcs(vcs, pretty = TRUE)
}
\arguments{
\item{vcs}{Named list of vc objects}

\item{pretty}{Logical. Pretty-print JSON.}
}
\value{
A JSON string
}
\description{
Export a list of variance components to JSON
}
