% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{as_kernel}
\alias{as_kernel}
\title{Coerce kernel specifications to kernel objects}
\usage{
as_kernel(x)
}
\arguments{
\item{x}{Kernel object or supported backend object}
}
\description{
Coerce kernel specifications to kernel objects
}
