% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.Datalist}
\alias{as_list.Datalist}
\title{Convert a Datalist object to a list for JSON export}
\usage{
as_list.Datalist(x)
}
\arguments{
\item{x}{Datalist object}
}
\value{
A named list
}
\description{
Convert a Datalist object to a list for JSON export
}
