% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.kernel}
\alias{as_list.kernel}
\title{Convert a kernel object to a list for JSON export}
\usage{
as_list.kernel(x)
}
\arguments{
\item{x}{Kernel object}
}
\value{
A named list
}
\description{
Convert a kernel object to a list for JSON export
}
