% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.kernels}
\alias{as_list.kernels}
\title{Convert kernel specifications to list form}
\usage{
as_list.kernels(vcs)
}
\arguments{
\item{vcs}{Named list of vc objects}
}
\value{
KernelSpec list
}
\description{
Convert kernel specifications to list form
}
