% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.model}
\alias{as_list.model}
\title{Convert model specification to list for JSON export}
\usage{
\method{as_list}{model}(formulas, task, options = list())
}
\arguments{
\item{formulas}{Named list of formulas}

\item{task}{Character ("reml", "bayes", "solve")}

\item{options}{Optional list}
}
\value{
Named list
}
\description{
Convert model specification to list for JSON export
}
