% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.prior}
\alias{as_list.prior}
\title{Convert a prior object to a list suitable for JSON export}
\usage{
as_list.prior(x)
}
\arguments{
\item{x}{An object of class "prior"}
}
\value{
Named list
}
\description{
Convert a prior object to a list suitable for JSON export
}
