% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.priors}
\alias{as_list.priors}
\title{Convert prior specifications to list form}
\usage{
as_list.priors(priors)
}
\arguments{
\item{priors}{Named list of prior objects}
}
\value{
PriorSpec list
}
\description{
Convert prior specifications to list form
}
