% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.vc}
\alias{as_list.vc}
\title{Convert a vc object to a list suitable for JSON export}
\usage{
as_list.vc(x)
}
\arguments{
\item{x}{An object of class "vc"}

\item{name}{Optional name of the variance component}
}
\value{
A named list
}
\description{
Convert a vc object to a list suitable for JSON export
}
