% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{as_list.vcs}
\alias{as_list.vcs}
\title{Convert variance component specifications to a list}
\usage{
as_list.vcs(vcs)
}
\arguments{
\item{vcs}{Named list of vc objects}
}
\value{
VarCompSpec list
}
\description{
Convert variance component specifications to a list
}
