% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_constructor.R
\name{bayesC}
\alias{bayesC}
\title{BayesC prior for marker effects}
\usage{
bayesC(pi, gamma, sets = NULL, weights = NULL, annotation = NULL)
}
\arguments{
\item{pi}{Mixing proportions (e.g. c(0.95, 0.05))}

\item{gamma}{Variance scaling parameters}

\item{sets}{Optional marker sets}

\item{weights}{Optional marker weights}

\item{annotation}{Optional functional annotation}
}
\value{
An object of class "bayesC"
}
\description{
BayesC prior for marker effects
}
