% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_constructor.R
\name{bayesR}
\alias{bayesR}
\title{BayesR prior for marker effects}
\usage{
bayesR(pi, variances, sets = NULL, weights = NULL, annotation = NULL)
}
\arguments{
\item{pi}{Mixing proportions}

\item{variances}{Vector of component variances}

\item{sets}{Optional marker sets}

\item{weights}{Optional marker weights}

\item{annotation}{Optional functional annotation}
}
\value{
An object of class "bayesR"
}
\description{
BayesR prior for marker effects
}
