% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{export_model_bundle}
\alias{export_model_bundle}
\title{Export full model specification bundle to JSON files}
\usage{
export_model_bundle(data, formulas, vcs, task, path = ".")
}
\arguments{
\item{data}{Datalist}

\item{formulas}{Model formulas}

\item{vcs}{Variance components}

\item{task}{Character}

\item{path}{Output directory}
}
\description{
Export full model specification bundle to JSON files
}
