% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{iid_kernel}
\alias{iid_kernel}
\title{IID kernel}
\usage{
iid_kernel(type = "iid")
}
\arguments{
\item{type}{Character. Label for IID kernel (e.g. "residual", "litter").}
}
\description{
IID kernel
}
