% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_constructor.R
\name{iw}
\alias{iw}
\title{Inverse-Wishart prior}
\usage{
iw(df, S)
}
\arguments{
\item{df}{Degrees of freedom}

\item{S}{Scale matrix}
}
\value{
An object of class "iw"
}
\description{
Inverse-Wishart prior
}
