% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{ld_kernel}
\alias{ld_kernel}
\title{LD-based kernel}
\usage{
ld_kernel(LDlist, sparse = TRUE)
}
\arguments{
\item{LDlist}{An LDlist object}

\item{sparse}{Logical; whether LD is sparse}
}
\description{
LD-based kernel
}
