% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_constructor.R
\name{new_prior_distribution}
\alias{new_prior_distribution}
\title{Base class for prior distributions}
\usage{
new_prior_distribution(type, params)
}
\description{
Base class for prior distributions
}
