% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_constructor.R
\name{prior_iw}
\alias{prior_iw}
\title{Inverse-Wishart prior for trait covariance}
\usage{
prior_iw(df, S)
}
\arguments{
\item{df}{Degrees of freedom.}

\item{S}{Scale matrix.}
}
\description{
Inverse-Wishart prior for trait covariance
}
