% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{qg_register_engine}
\alias{qg_register_engine}
\title{Register a computational engine}
\usage{
qg_register_engine(name, fun)
}
\arguments{
\item{name}{Engine name.}

\item{fun}{A function with signature \code{function(compiled, ...)} returning a fit object.}
}
\description{
Register a computational engine
}
