% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{qgtools_set_threads}
\alias{qgtools_set_threads}
\title{Set number of threads for qgtools}
\usage{
qgtools_set_threads(n)
}
\arguments{
\item{n}{number of threads}
}
\description{
Set number of threads for qgtools
}
