% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{qgtools_threads_info}
\alias{qgtools_threads_info}
\title{Show threading configuration for qgtools}
\usage{
qgtools_threads_info()
}
\description{
Prints information about detected cores and thread settings
for OpenMP and BLAS backends.
}
