% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{read_qg_yaml}
\alias{read_qg_yaml}
\title{Read a qgtools YAML specification}
\usage{
read_qg_yaml(path)
}
\arguments{
\item{path}{Path to a YAML file}
}
\value{
A list representing the qgtools model specification
}
\description{
Read a qgtools YAML specification
}
