% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_exporters.R
\name{validate_bundle_json}
\alias{validate_bundle_json}
\title{Validate consistency of a model specification bundle}
\usage{
validate_bundle_json(
  data_spec,
  model_spec,
  kernel_spec,
  varcomp_spec = NULL,
  prior_spec = NULL
)
}
\arguments{
\item{data_spec}{DataSpec list}

\item{model_spec}{ModelSpec list}

\item{kernel_spec}{KernelSpec list}

\item{varcomp_spec}{Optional VarCompSpec list}

\item{prior_spec}{Optional PriorSpec list}
}
\description{
Ensures that DataSpec, ModelSpec, KernelSpec, and either VarCompSpec
(REML / solver) or PriorSpec (Bayesian) are mutually consistent.
}
