% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_bundle_reml}
\alias{validate_bundle_reml}
\title{Validate consistency of a model specification bundle}
\usage{
validate_bundle_reml(data_spec, model_spec, varcomp_spec, kernel_spec)
}
\arguments{
\item{data_spec}{DataSpec list (from as_list.Datalist)}

\item{model_spec}{ModelSpec list (from as_list.model)}

\item{varcomp_spec}{VarCompSpec list (from as_list.vcs)}

\item{kernel_spec}{KernelSpec list (from as_list.kernels)}
}
\description{
Validate consistency of a model specification bundle
}
