% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructor.R, R/validators.R
\name{validate_kernel}
\alias{validate_kernel}
\title{Validate a single kernel object}
\usage{
validate_kernel(kernel, kernel_id)

validate_kernel(kernel, kernel_id)
}
\description{
Validate a single kernel object

Validate a kernel object based on inferred type
}
\keyword{internal}
