% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klist.R
\name{validate_Klist}
\alias{validate_Klist}
\title{Validate a Klist}
\usage{
validate_Klist(Klist, strict = TRUE)
}
\arguments{
\item{Klist}{A Klist or named list of kernels}

\item{strict}{If TRUE, error; else return character vector of issues.}
}
\description{
Validate a Klist
}
\keyword{internal}
