% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_qg_model}
\alias{validate_qg_model}
\title{Validate qgtools model specification}
\usage{
validate_qg_model(models, vcov = NULL, Klist = NULL)
}
\description{
Validate qgtools model specification
}
\keyword{internal}
