% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_vcov_sets_kernel}
\alias{validate_vcov_sets_kernel}
\title{Validate vcov sets vs kernel compatibility}
\usage{
validate_vcov_sets_kernel(vcov, Klist)
}
\description{
Validate vcov sets vs kernel compatibility
}
\keyword{internal}
