% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_vcov_trait_structure}
\alias{validate_vcov_trait_structure}
\title{Validate vcov trait-structure compatibility}
\usage{
validate_vcov_trait_structure(vcov)
}
\description{
Validate vcov trait-structure compatibility
}
\keyword{internal}
