% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructor.R
\name{validateGRMlist}
\alias{validateGRMlist}
\title{Validate a GRMlist object}
\usage{
validateGRMlist(x, check.file = TRUE, check.format = TRUE)
}
\arguments{
\item{x}{An object of class \code{"GRMlist"}.}

\item{check.file}{Logical. Check that the GRM file exists.}

\item{check.format}{Logical. Lightweight structure check (TEXT only).}
}
\value{
Invisibly returns TRUE if validation succeeds.
}
\description{
Perform consistency and integrity checks on a \code{GRMlist} object.
}
