% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructor.R
\name{validateLDlist}
\alias{validateLDlist}
\title{Validate an LDlist object}
\usage{
validateLDlist(x, check.file = TRUE, check.format = TRUE)
}
\arguments{
\item{x}{An object of class \code{"LDlist"}.}

\item{check.file}{Logical. Check that the LD file exists.}

\item{check.format}{Logical. Lightweight structure check (TEXT only).}
}
\value{
Invisibly returns TRUE if validation succeeds.
}
\description{
Perform consistency and integrity checks on an \code{LDlist} object.
}
