% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructor.R
\name{validatePEDlist}
\alias{validatePEDlist}
\title{Validate a PEDlist object}
\usage{
validatePEDlist(x, check.file = TRUE, check.format = TRUE, check.logic = FALSE)
}
\arguments{
\item{x}{An object of class \code{"PEDlist"}.}

\item{check.file}{Logical. Check that the pedigree file exists.}

\item{check.format}{Logical. Lightweight structure check (TEXT only).}

\item{check.logic}{Logical. Optional deep pedigree logic checks (default FALSE).}
}
\description{
Perform consistency and integrity checks on a \code{PEDlist} object.
}
