% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{varcomp}
\alias{varcomp}
\title{Variance component collection}
\usage{
varcomp(x = NULL, ...)
}
\arguments{
\item{...}{Named variance component objects of class "vc".}
}
\value{
An object of class "varcomp".
}
\description{
Collect and validate a set of variance component specifications.
}
