% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{write_qg_yaml}
\alias{write_qg_yaml}
\title{Write qgtools model specification to YAML}
\usage{
write_qg_yaml(formulas, vcov = NULL, Klist = NULL, file)
}
\arguments{
\item{formulas}{Named list of formulas, one per trait}

\item{vcov}{List specifying trait-level covariance structures}

\item{Klist}{Optional Klist object (used only for kernel names)}

\item{file}{Output YAML filename}
}
\value{
Invisibly returns the YAML object
}
\description{
Write qgtools model specification to YAML
}
