% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{grm_kernel}
\alias{grm_kernel}
\title{Genomic relationship matrix (GRM) kernel}
\usage{
grm_kernel(GRMlist)
}
\arguments{
\item{GRMlist}{A GRMlist object}
}
\description{
Genomic relationship matrix (GRM) kernel
}
