% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{iid_kernel}
\alias{iid_kernel}
\title{IID kernel (independent and identically distributed)}
\usage{
iid_kernel(label = "iid")
}
\arguments{
\item{label}{Optional label (e.g. "residual", "dam")}
}
\description{
Used for random intercepts, residuals, litter effects, etc.
}
