% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{ld_kernel}
\alias{ld_kernel}
\title{LD-based marker kernel}
\usage{
ld_kernel(LDlist, sparse = TRUE)
}
\arguments{
\item{sparse}{Logical. Whether LD is sparse.}

\item{Glist}{A Glist object.}
}
\description{
LD-based marker kernel
}
