% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_constructor.R
\name{makePEDlist}
\alias{makePEDlist}
\title{Create a PEDlist object}
\usage{
makePEDlist(fnPED = NULL, format = "ASCII", method = "S-D-NonInbred")
}
\description{
Create a pedigree data specification object (\code{PEDlist}) used to define
the structure of additive genetic relationships based on pedigree information.
}
\details{
The \code{PEDlist} object contains no phenotype or model-specific information.
Linking pedigree records to analysis data is handled at a higher (model) level.
}
