% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgtools-package.R
\docType{package}
\name{qgtools-package}
\alias{qgtools}
\alias{qgtools-package}
\title{qgtools: Generic Tools for Linear (Mixed) Models}
\description{
Infrastructure package for specifying linear (mixed) models with
genomic, pedigree, and multi-kernel covariance structures, decoupled
from computational backends.
}
\author{
\strong{Maintainer}: Peter Soerensen \email{first.last@example.com}

}
\keyword{internal}
