% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_kernel}
\alias{validate_kernel}
\title{Validate a kernel object based on inferred type}
\usage{
validate_kernel(kernel, kernel_id)
}
\description{
Validate a kernel object based on inferred type
}
\keyword{internal}
