% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_component_constructor.R
\name{varcomp}
\alias{varcomp}
\title{Variance component collection}
\usage{
varcomp(x = NULL, ...)
}
\arguments{
\item{...}{Named variance component objects of class \code{"vc"}.}
}
\value{
An object of class \code{"varcomp"}.
}
\description{
Collect and validate a set of variance component specifications.
Each component must be created with \code{vc()}.
}
